Change time_to_live from int to float everywhere

+++ src/state/include/actor/arrow.h
@@ -35,14 +35,14 @@ private:
 	 * Lifetime of an arrow
 	 * Dies after `time_to_live` clock ticks
 	 */
-	int64_t time_to_live;
+	float time_to_live;
 public:
 	Arrow(
 		act_id_t id,
 		PlayerId player_id,
 		int64_t attack,
 		int64_t max_speed,
-		int64_t time_to_live,
+		float time_to_live,
 		int64_t size,
 		physics::Vector2D start,
 		Actor* target
